+++ POI/trunk/org.openl.lib.poi35.dev.modified/src/org/apache/poi/hssf/record/formula/functions/XYNumericFunction.java
@@ -121,7 +121,7 @@ public abstract class XYNumericFunction implements FunctionWithArraySupport {
 	 */
 	protected abstract Accumulator createAccumulator();
 
-	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 		if (args.length != 2) {
 			return ErrorEval.VALUE_INVALID;
 		}
